refactor(progress): break up into multiple modules
